style: apply forge fmt formatting

+++ script/HelperConfig.s.sol
@@ -4,7 +4,7 @@ import {Script} from "forge-std/Script.sol";
 import {VRFCoordinatorV2_5Mock} from "chainlink-evm/contracts/src/v0.8/vrf/mocks/VRFCoordinatorV2_5Mock.sol";
 
 abstract contract CodeConstants {
-    uint public MOCK_BASE_FEE = 0.25 ether;
+    uint256 public MOCK_BASE_FEE = 0.25 ether;
     uint256 public MOCK_GAS_PRICE_LINK = 1e9; // 0.000000001 LINK per gas
     int256 public MOCK_WEI_PER_UNIT_LINK = 4e15; // 1
     uint256 public constant SEPOLIA_CHAIN_ID = 11155111;
@@ -13,6 +13,7 @@ abstract contract CodeConstants {
 
 contract HelperConfig is CodeConstants, Script {
     error HelperConfig__invalidChainId();
+
     struct NetworkConfig {
         uint256 entranceFee;
         uint256 interval;
@@ -28,9 +29,7 @@ contract HelperConfig is CodeConstants, Script {
         networkConfigs[SEPOLIA_CHAIN_ID] = getSepoliaEthConfig();
     }
 
-    function getConfigByChainId(
-        uint256 chainId
-    ) public returns (NetworkConfig memory) {
+    function getConfigByChainId(uint256 chainId) public returns (NetworkConfig memory) {
         if (networkConfigs[chainId].vrfCoordinator != address(0)) {
             return networkConfigs[chainId];
         } else if (chainId == LOCALHOST_CHAIN_ID) {
@@ -45,15 +44,14 @@ contract HelperConfig is CodeConstants, Script {
     }
 
     function getSepoliaEthConfig() public pure returns (NetworkConfig memory) {
-        return
-            NetworkConfig({
-                entranceFee: 0.01 ether,
-                interval: 30,
-                vrfCoordinator: 0x9DdfaCa8183c41ad55329BdeeD9F6A8d53168B1B,
-                keyHash: 0x787d74caea10b2b357790d5b5247c2f63d1d91572a9846f780606e4d953677ae,
-                subscriptionId: 0,
-                callbackGasLimit: 500000
-            });
+        return NetworkConfig({
+            entranceFee: 0.01 ether,
+            interval: 30,
+            vrfCoordinator: 0x9DdfaCa8183c41ad55329BdeeD9F6A8d53168B1B,
+            keyHash: 0x787d74caea10b2b357790d5b5247c2f63d1d91572a9846f780606e4d953677ae,
+            subscriptionId: 0,
+            callbackGasLimit: 500000
+        });
     }
 
     function getCreateAnvilEthConfig() public returns (NetworkConfig memory) {
@@ -61,11 +59,8 @@ contract HelperConfig is CodeConstants, Script {
             return localNetworkConfig;
         } else {
             vm.startBroadcast();
-            VRFCoordinatorV2_5Mock vrfCoordinator = new VRFCoordinatorV2_5Mock(
-                MOCK_BASE_FEE,
-                MOCK_GAS_PRICE_LINK,
-                MOCK_WEI_PER_UNIT_LINK
-            );
+            VRFCoordinatorV2_5Mock vrfCoordinator =
+                new VRFCoordinatorV2_5Mock(MOCK_BASE_FEE, MOCK_GAS_PRICE_LINK, MOCK_WEI_PER_UNIT_LINK);
             vm.stopBroadcast();
             localNetworkConfig = NetworkConfig({
                 entranceFee: 0.01 ether,
